Update for index option
1/ 2 exceptions in case of keys [res_name+" "+atom_name] not found in param files. Add those 2 exceptions and errors in m.get_radius and m.get_aliphatic. A if checks if the association res_name and atom_name is found in the param file. If not, it send an error message and exit the program. 

2/ 
a) I add an option (-n) to take into account and index for upper/lower leaflet. Useful in case of undulation where the Z coord is not enough to determine the two leaflets. The index file is in ndx format (Gromacs) and should only have lower and upper groups. 
Exemple :
[ Lower ]
 3 4 5
[ Upper ] 
6 7 8  
It works also for [ lower ] and [ upper ] or any name including the words "Lower"/"lower" and "Upper"/"upper". Be careful, if lower or upper words is found, the program will assume that the other group is the opposite leaflet. 

b) I add a function in param.py named read_ndx(ndx_file) to read and set the lower_leaflet and upper_leaflet variables. I used the bfrg.read_file option tha